email-intelligence/backend/requirements.txt
fastapi
uvicorn
python-dotenv

psycopg2-binary
sqlalchemy

qdrant-client

google-api-python-client
google-auth
google-auth-oauthlib

pydantic

pydantic-settings
